MyDirect3D12Engine v0.3

+++ MyDirect3D12Engine/MyDirect3D12Engine/BasicGeometryGenerator.h
@@ -13,6 +13,6 @@ private:
 	std::vector<Vertex> mVertices;
 	UINT mVertexCount;
 
-	std::vector<uint16_t> mIndices;
+	std::vector<uint32_t> mIndices;
 	UINT mIndexCount;
 };
